+++ RiskOfChaos/Utilities/ParsedValueHolders/ParseException.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace RiskOfChaos.Utilities.ParsedValueHolders
+{
+    public sealed class ParseException : Exception
+    {
+        public ParseException(string message) : base(message)
+        {
+        }
+    }
+}