+++ RiskOfChaos/Utilities/ParsedValueHolders/ParseFailReason.cs
@@ -4,5 +4,5 @@ using System.Text;
 
 namespace RiskOfChaos.Utilities.ParsedValueHolders
 {
-    public record class ParseFailReason(string ParseInput, Exception ParseException);
+    public record class ParseFailReason(string ParseInput, ParseException ParseException);
 }